pyproject.toml
[tool.poetry]
name = "poolboy"
version = "0.1.0"
description = ""
authors = ["MrTrustworthy <tinywritingmakesmailhardtoread@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.5"
djangorestframework = "^3.11"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
